feat: update to dart v3
