+++ pubspec.yaml
@@ -6,11 +6,10 @@ repository: https://github.com/Aallam/ulid4d
 issue_tracker: https://github.com/Aallam/ulid4d/issues
 
 environment:
-  sdk: '>=2.17.6 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   fixnum: ^1.0.1
-  meta: ^1.8.0
 
 dev_dependencies:
   benchmark_harness: ^2.2.0
