+++ test/ulid_test.dart
@@ -1,6 +1,7 @@
 import 'package:fixnum/fixnum.dart';
 import 'package:test/test.dart';
 import 'package:ulid4d/src/ulid.dart';
+import 'package:ulid4d/src/ulid_default.dart';
 import 'package:ulid4d/src/ulid_factory.dart';
 import 'package:ulid4d/src/utils.dart';
